be/pyproject.toml
[project]
name = "orbi"
version = "0.1.0"
description = "Audio upload API with FastAPI and Speaker Intelligence"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # Async Task Queue
    "celery>=5.3.0",
    "redis>=5.0.0",

    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",

    # Audio Processing
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "pydub>=0.25.0",
    "audioread>=3.0.0",

    # Speaker Diarization & Recognition
    "pyannote.audio>=3.1.0",
    "speechbrain>=0.5.0",

    # Machine Learning
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "transformers>=4.30.0",

    # Transcription
    "openai-whisper>=20231117",

    # Vector Database
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",

    # LLM Integration
    "openai>=1.0.0",
    "anthropic>=0.7.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.24.0",
    "faker>=19.0.0",
]
